+++ src/app/state/reducers.ts
@@ -5,9 +5,11 @@ import {
 import { environment } from '../../environments/environment';
 import { settingsReducer } from './settings/settings.reducer';
 import {State} from "./state";
+import { categoriesReducer } from './categories/categories.reducer';
 
 export const reducers : ActionReducerMap<State> = {
   settings: settingsReducer,
+  categories: categoriesReducer,
 };