+++ src/app/state/selectors.ts
@@ -2,8 +2,10 @@ import {Settings} from "../models/settings.model";
 import { createSelector } from '@ngrx/store';
 import * as dayjs from 'dayjs';
 import { Days } from '../models/overallLeft.model';
+import { Categories } from '../models/categories.model';
 
 export const selectSettings = (state: any): Settings => state.settings;
+export const selectCategories = (state: any): Categories => state.categories;
 
 export const selectDays = createSelector(
   selectSettings,