+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - settings type
 - [NGRX](https://ngrx.io/)
 - [ESLint](https://eslint.org/)
+- categories storage
 
 ## [3.2.0] - 2021-12-24
 ### Added